Crate native_windows_gui

A very high level native gui library for Windows.

Modules

keys
Windows virtual key code

Structs

Menu
A menu control

MenuItem
A menu item control

MenuItemT
A template to create menuitems

MenuT
A template to create menu controls

Ui
Object that manage the GUI elements

Window
A window control.

WindowT
A template that will create a window.

Enums

AnyHandle
A type that expose the different underlying handle into one type

Error
Error class that regroup errors generated by NWG

Event
Events name definition

EventArgs
Events arguments definition. If an event do not have arguments, EventArgs::None is passed.

SystemError
Error class that regroups errors generated by the system

Traits

Control
Structures implementing this trait are controls that can be stored in a Ui

ControlT
Structures implementing this trait can be used by a Ui to build a Control

Functions

dispatch_events
Dispatch the messages waiting the the system message queue to the associated Uis. This includes NWG custom messages.

exit
Send a WM_QUIT to the system queue. Breaks the dispatch_events loop.

Type Definitions

EventCallback
Event callback type