Module native_windows_gui::constants

Controls constants

Modules

keys
Windows virtual key code

Structs

Ellipse
Define a ellipse shape that can be used with canvases

MessageParams
A structure that defines how a messagebox should look and behave.

Pen
Describe how lines should be painted. Used when painting in a canvas

PickerDate
A date struct that can be passed to a date time picker control.

Rectangle
Define a rectangle shape that can be used with canvases

SYSTEMTIME

SolidBrush
A brush using a single solid color. Used when painting in a canvas

Enums

CapStyle
Cap style used when creating a Pen

CheckState
Checkbox checkstate

ControlType
An enum that list type names for the common controls.

DashStyle
Dash style used when creating a Pen

FileDialogAction
A enum that dictates how a file dialog should behave

HTextAlign
Text align constant on the horizontal axis

LineJoin
Line join type used when creating a Pen

MessageButtons
Enum of message box buttons (to use with MessageParams )

MessageChoice
Return value of message. Define the button that the user clicked. If the user cancelled the message box by clicking on X button of the window, MessageChoice::Cancel is returned.

MessageIcons
Enum of message box icons (to use with MessageParams )

MouseButton
Mouse buttons

ProgressBarState
Define the state of a progress bar

VTextAlign
Text align constant on the vertical axis

Constants

CENTER_POSITION

DEFAULT_POSITION

DEFAULT_SIZE

FONT_DECO_ITALIC

FONT_DECO_NORMAL

FONT_DECO_STRIKEOUT

FONT_DECO_UNDERLINE

FONT_WEIGHT_BLACK

FONT_WEIGHT_BOLD

FONT_WEIGHT_DONTCARE

FONT_WEIGHT_EXTRABOLD

FONT_WEIGHT_EXTRALIGHT

FONT_WEIGHT_LIGHT

FONT_WEIGHT_MEDIUM

FONT_WEIGHT_NORMAL

FONT_WEIGHT_SEMIBOLD

FONT_WEIGHT_THIN