Module native_windows_gui::custom

Custom control creation resources

Structs

SysclassParams
Params used to build a system class

WindowParams
Params used to build a system window

Enums

AnyHandle
A type that expose the different underlying handle into one type

Traits

Control
Structures implementing this trait are controls that can be stored in a Ui

ControlT
Structures implementing this trait can be used by a Ui to build a Control

Resource
Structures implementing this trait are resources that can be stored in a Ui

ResourceT
Structures implementing this trait can be used by a Ui to build a Resource

Functions

build_sysclass⚠
Try to create a system class using the parameters provided in SysclassParams. Will not fail if the system class already exists.

build_window⚠
Try to create a system class using the parameters provided in WindowParams.

get_window_enabled⚠
Get the window enabled state

get_window_long

get_window_position⚠
Get window position

get_window_size⚠
Get window size

get_window_text⚠
Get the window text

get_window_visibility⚠
Get window visibility

set_window_enabled⚠
Set the window enabled state

set_window_long

set_window_position⚠
Set window position

set_window_size⚠
Set window size

set_window_text⚠
Set the window text

set_window_visibility⚠
Set window visibility