Enum native_windows_gui::custom::Event

pub enum Event {
    Any,
    Single(UINT, &'static UnpackProc, &'static HandleProc),
    Group(&'static [UINT], &'static UnpackProc, &'static HandleProc),
}

An enum that define events that can be used by NWG

Variants
Any

A special identifier that catches every system messages

Single(UINT, &'static UnpackProc, &'static HandleProc)

Wrap a single system message identified by the first paramater

Group(&'static [UINT], &'static UnpackProc, &'static HandleProc)

Wrap a group of system messages identified by the first paramater