Function native_windows_gui::custom::build_window

pub unsafe fn build_window<S1: Into<String>, S2: Into<String>>(
    p: WindowParams<S1, S2>
) -> Result<HWND, SystemError>

Try to create a system class using the parameters provided in WindowParams.

Returns Ok(HWND) where HWND is the newly created window handle Returns Err(SystemError::WindowCreationFail) if the system window creation failed.

Note that if the system class window proc used is malformed, the program will most likely segfault.