Trait native_windows_gui::custom::Control

pub trait Control {
    fn handle(&self) -> AnyHandle;

    fn control_type(&self) -> ControlType { ... }
    fn free(&mut self) { ... }
}

Structures implementing this trait are controls that can be stored in a Ui

Required Methods

fn handle(&self) -> AnyHandle

Should return the underlying handle to the object

Provided Methods

fn control_type(&self) -> ControlType

Should return the type of the control. For custom controls, the return value should be Undefined (the default).

fn free(&mut self)

If specified, should free any ressource allocated in the template build function. This includes functions like DestroyWindow.