Trait native_windows_gui::custom::ControlT

pub trait ControlT<ID: Clone + Hash> {
    fn type_id(&self) -> TypeId;
    fn build(&self, ui: &Ui<ID>) -> Result<Box<Control>, Error>;

    fn events(&self) -> Vec<Event> { ... }
}

Structures implementing this trait can be used by a Ui to build a Control

Required Methods

fn type_id(&self) -> TypeId

Should return the TypeId of the generated control. For example a WindowT struct returns the TypeId of a Window struct.

fn build(&self, ui: &Ui<ID>) -> Result<Box<Control>, Error>

Should instance the control and return it as a Box. If an error is raised, it will be returned by ui.commit.

Provided Methods

fn events(&self) -> Vec<Event>

Should return the events supported by the control.