Trait native_windows_gui::custom::Resource

pub trait Resource {
    fn handle(&self) -> AnyHandle;

    fn free(&mut self) { ... }
}

Structures implementing this trait are resources that can be stored in a Ui

Required Methods

fn handle(&self) -> AnyHandle

Should return the underlying handle to the object

Provided Methods

fn free(&mut self)

If specified, should free any ressource allocated in the template build function.