Struct native_windows_gui::CanvasRenderer

pub struct CanvasRenderer<'a, ID: Clone + Hash + 'a> { /* fields omitted */ }

Object that offers a light wrapper over the D2D1 api.

Methods

impl<'a, ID: Clone + Hash> CanvasRenderer<'a, ID>

fn clear(&mut self, r: f32, g: f32, b: f32, a: f32)

Clears the drawing area to the specified color.

Arguments:
• r: Red component. 0.0 - 1.0
• g: Green component. 0.0 - 1.0
• b: Blue component. 0.0 - 1.0
• a: Alpha component. 0.0 - 1.0

fn get_render_size(&mut self) -> (f32, f32)

Return the render target size in a tuple of (width, height)

fn set_transform(&mut self, m: &[[FLOAT; 2]; 3])

Set the transformation matrix of the renderer

Arguments:
• m: The 3x2 matrix.

fn get_transform(&mut self) -> [[FLOAT; 2]; 3]

Return the transformation matrix of the renderer

fn fill_rectangle(&mut self, brush: &ID, r: &Rectangle) -> Result<(), Error>

Fill a rectangle shape defined by r using the brush identified by brush

Arguments:
• brush: Id of the brush saved in the canvas
• r: The rectangle shape to draw

fn fill_rounded_rectangle(
&mut self,
brush: &ID,
r: &Rectangle,
radius: (f32, f32)
) -> Result<(), Error>

Fill a rounded rectangle shape defined by r using the brush identified by brush

Arguments:
• brush: Id of the brush saved in the canvas
• r: The rectangle shape to draw
• radius: Amount of rounding on the rectangle border. (width, height)

fn fill_ellipse(&mut self, brush: &ID, e: &Ellipse) -> Result<(), Error>

Fill an ellipse shape defined by e using the brush identified by brush

Arguments:
• brush: Id of the brush saved in the canvas
• e: The ellipse shape to draw

fn draw_rectangle(
&mut self,
brush: &ID,
pen: Option<&ID>,
r: &Rectangle,
width: f32
) -> Result<(), Error>

Draw the outline of a rectangle shape defined by r using the brush identified by brush and the pen identifed by pen.

Arguments:
• brush: Id of the brush saved in the canvas
• pen: Id of the pen saved in the canvas
• r: The rectangle shape to draw
• width: Width of the outline to draw

fn draw_rounded_rectangle(
&mut self,
brush: &ID,
pen: Option<&ID>,
r: &Rectangle,
width: f32,
radius: (f32, f32)
) -> Result<(), Error>

Draw the outline of a rounded rectangle shape defined by r using the brush identified by brush and the pen identifed by pen.

Arguments:
• brush: Id of the brush saved in the canvas
• pen: Id of the pen saved in the canvas
• r: The rectangle shape to draw
• width: Width of the outline to draw
• radius: Amount of rounding on the rectangle border. (width, height)

fn draw_ellipse(
&mut self,
brush: &ID,
pen: Option<&ID>,
e: &Ellipse,
width: f32
) -> Result<(), Error>

Draw the outline of an ellipse shape defined by e using the brush identified by brush and the pen identifed by pen.

Arguments:
• brush: Id of the brush saved in the canvas
• pen: Id of the pen saved in the canvas
• e: The ellipse shape to draw
• width: Width of the outline to draw