Struct native_windows_gui::ListBoxT

pub struct ListBoxT<D: Clone + Display + 'static, ID: Hash + Clone> {
    pub collection: Vec<D>,
    pub position: (i32, i32),
    pub size: (u32, u32),
    pub visible: bool,
    pub disabled: bool,
    pub readonly: bool,
    pub multi_select: bool,
    pub parent: ID,
    pub font: Option<ID>,
}

Template that creates a listbox control

Available events:
Destroyed, Moved, Resized, SelectionChanged, DoubleClick, Focus, Any

Members:
• collection: Item collection of the listbox. The item type must implement Display
• position: The start position of the listbox
• size: The start size of the listbox
• visible: If the listbox should be visible to the user
• disabled: If the user can or can't use the listbox
• readonly : If true, the user won't be able to select items in the listbox
• multi_select: If true, allow the user to select more than one item
• parent: The listbox parent
• font: The listbox font. If None, use the system default

Fields
collection: Vec<D> position: (i32, i32) size: (u32, u32) visible: bool disabled: bool readonly: bool multi_select: bool parent: ID font: Option<ID>