Struct native_windows_gui::ProgressBar

pub struct ProgressBar { /* fields omitted */ }

A standard progress bar

Methods

impl ProgressBar

fn get_range(&self) -> (u32, u32)

Return the current range of the progress bar

fn set_range(&self, min: u32, max: u32) -> Result<(), Error>

Set the extent of the progress bar control. If min is bigger than max, an error will be returned.

fn get_step(&self) -> u32

Return the current step of the progress bar

fn set_step(&self, step: u32)

Set the step of the progress bar control. When step() is called, the progress bar value is increased by the value defined by step.

fn get_value(&self) -> u32

Return the current value of the progress bar

fn set_value(&self, val: u32)

Set the progress bar value

fn get_state(&self) -> ProgressBarState

Get the progress bar state

fn set_state(&self, state: ProgressBarState)

Set the progress bar state

fn step(&self)

fn advance(&self, amount: u32)

Add amount to the progress bar value

fn get_visibility(&self) -> bool

fn set_visibility(&self, visible: bool)

fn get_position(&self) -> (i32, i32)

fn set_position(&self, x: i32, y: i32)

fn get_size(&self) -> (u32, u32)

fn set_size(&self, w: u32, h: u32)

fn get_enabled(&self) -> bool

fn set_enabled(&self, e: bool)