Struct native_windows_gui::ProgressBarT

pub struct ProgressBarT<ID: Hash + Clone> {
    pub position: (i32, i32),
    pub size: (u32, u32),
    pub visible: bool,
    pub disabled: bool,
    pub range: (u32, u32),
    pub value: u32,
    pub step: u32,
    pub state: ProgressBarState,
    pub vertical: bool,
    pub parent: ID,
}

A template that creates a progress bar

Events:
Destroyed, Moved, Resized, Any

Members:
• position: The start position of the progressbar
• size: The start size of the progressbar
• visible: If the progressbar should be visible to the user
• disabled: If the user can or can't click on the progressbar
• range: The minimum and the maximum value of the progress bar
• position: The starting position of the progress bar
• step: Amount of value to add when step is called
• state: The state of the progress bar. • vertical: If the progress bar should be vertical instead of horizontal • parent: The progressbar parent

Fields
position: (i32, i32) size: (u32, u32) visible: bool disabled: bool range: (u32, u32) value: u32 step: u32 state: ProgressBarState vertical: bool parent: ID