Struct native_windows_gui::WindowT

pub struct WindowT<S: Clone + Into<String>> {
    pub title: S,
    pub position: (i32, i32),
    pub size: (u32, u32),
    pub resizable: bool,
    pub visible: bool,
    pub disabled: bool,
    pub exit_on_close: bool,
}

A template that will create a window.

Events:
Destroyed, KeyDown, KeyUp, Char, Closed, MouseDown, MouseUp, Moved, Resized, Any

Members:
• title : The title of the window (in the title bar)
• position : Starting posiion of the window after it is created
• size : Starting size of the window after it is created
• resizable : If the user can resize the window or not
• visible : If the user can see the window or not
• disabled : If the window is enabled or not. A disabled window do not process events
• exit_on_close : If NWG should break the event processing loop when this window is closed

Fields
title: S position: (i32, i32) size: (u32, u32) resizable: bool visible: bool disabled: bool exit_on_close: bool