Module native_windows_gui::custom

Custom control creation resources

Structs

SysclassParams
Params used to build a system class

WindowParams
Params used to build a system window

Enums

AnyHandle
A type that expose the different underlying handle into one type

Event
An enum that define events that can be used by NWG

Traits

Control
Structures implementing this trait are controls that can be stored in a Ui

ControlT
Structures implementing this trait can be used by a Ui to build a Control

Resource
Structures implementing this trait are resources that can be stored in a Ui

ResourceT
Structures implementing this trait can be used by a Ui to build a Resource

Functions

build_sysclass⚠
Try to create a system class using the parameters provided in SysclassParams. Will not fail if the system class already exists.

build_window⚠
Try to create a system class using the parameters provided in WindowParams.

command_handle
HandleProc for events using a WM_COMMAND message. Will return None if cmd do not match the WM_COMMAND code Should be used in a closure like this &|h,m,w,l|{ command_handle(h,m,w,l,BN_CLICKED) }

event_unpack_no_args
UnpackProc for events that have no arguments

get_window_enabled⚠
Get the window enabled state

get_window_long

get_window_position⚠
Get window position

get_window_size⚠
Get window size

get_window_text⚠
Get the window text

get_window_visibility⚠
Get window visibility

hwnd_handle
HandleProc for events that simply wrap the hwnd parameter

notify_handle
HandleProc for events using a WM_NOTIFY message. Will return None if cmd do not match the WM_COMMAND code

set_window_enabled⚠
Set the window enabled state

set_window_long

set_window_position⚠
Set window position

set_window_size⚠
Set window size

set_window_text⚠
Set the window text

set_window_visibility⚠
Set window visibility

Type Definitions

HandleProc
A procedure signature that takes raw message parameters and output a Handle Can return None if the handle could not be parsed

UnpackProc
A procedure signature that takes raw message parameters and output a EventArgs structure. Can return None if the parameters could not be parsed